Fix tests
Encountered 2 issues when running the tests on the master branch.

We used the be_true Rspec matcher, that was renamed to be_truthy in
Rspec ≥3.

Starting from Rspec 3, pending specs are expected to fail because they
represent something that cannot be tested due to currently failing code.

We had a pending spec without any code inside of it; I updated it to
skip instead of pending.